web_crawler/requirements.txt
beautifulsoup4==4.6.0
tabulate-0.8.2
